requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.0
torchvision

easydict
gradio
numpy
opencv-python
pillow

